select_device_from_magicgui_combobox
